💄 네비바 스타일 변경
